+++ lizrd/support/misc.py
@@ -3,8 +3,6 @@ import hashlib
 import random
 import string
 from typing import Optional, List, Tuple, Set
-import numpy as np
-import torch
 
 import yaml
 
@@ -83,6 +81,9 @@ def get_yaml_md5(file_path):
 
 
 def set_seed(seed):
+    import numpy as np
+    import torch
+
     torch.manual_seed(seed)
     torch.cuda.manual_seed_all(seed)  # if you are using multi-GPU.
     np.random.seed(seed)  # Numpy module.
@@ -90,5 +91,7 @@ def set_seed(seed):
 
 
 def get_ith_chunk(tensor, chunks, i):
+    import torch
+
     list_of_chunks = torch.chunk(tensor, chunks, dim=0)
     return list_of_chunks[i]
