+++ research/blanks/configs/scratchpad/example.yaml
