+++ research/blanks/configs/scratchpad/less_embeddings.yaml
